+++ matchms/importing/load_from_mgf.py
@@ -11,7 +11,7 @@ def load_from_mgf(filename: str) -> Generator[Spectrum, None, None]:
 
     .. code-block:: python
 
-        from matchs.importing import load_from_mgf
+        from matchms.importing import load_from_mgf
 
         file_mgf = "pesticides.mgf"
         spectrums = list(load_from_mgf(file_mgf))