+++ matchms/networking/__init__.py
@@ -1,0 +1,11 @@
+"""
+Functions for creating and analysing spectral networks
+######################################################
+
+"""
+from .SimilarityNetwork import SimilarityNetwork
+
+
+__all__ = [
+    "SimilarityNetwork",
+]